requirements.txt
# Core
streamlit==1.40.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15

# LLM
requests>=2.31.0

# Memory & Vector Store
chromadb==0.4.22
sentence-transformers==2.2.2

# NLP Utilities
textblob==0.17.1

# Database
SQLAlchemy==2.0.23

# Environment
python-dotenv==1.0.0

# Plotting
plotly==5.22.0

# Testing
pytest==7.4.3
